+++ docs/changelog.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [1.4.1] - 2024-09-02 {: #v1.4.1 }
+
 ### Added
 
 - An extra argument was added to `convert_markdown` to specify CSS files.
@@ -133,6 +136,7 @@ Initial release
 
 
 [Unreleased]: https://github.com/reproducible-reporting/stepup-reprep
+[1.4.1]: https://github.com/reproducible-reporting/stepup-reprep/releases/tag/v1.4.1
 [1.4.0]: https://github.com/reproducible-reporting/stepup-reprep/releases/tag/v1.4.0
 [1.3.0]: https://github.com/reproducible-reporting/stepup-reprep/releases/tag/v1.3.0
 [1.2.1]: https://github.com/reproducible-reporting/stepup-reprep/releases/tag/v1.2.1